// frontend/types.ts
export interface StoreSettings {
  storeNameBn: string;
  storeNameEn: string;
  taglineBn?: string;
  taglineEn?: string;
  phonePrimary: string;
  phoneSecondary?: string;
  email: string;
  addressBn: string;
  addressEn?: string;
  deliveryChargeInsideDhaka: number;
  deliveryChargeOutsideDhaka: number;
  freeDeliveryThreshold: number;
  announcementBn?: string;
  announcementEn?: string;
  announcementTextBn?: string;
  announcementTextEn?: string;
  announcementActive?: boolean;
  currencySymbol?: string;
  bkashNumber?: string;
  nagadNumber?: string;
  heroBackgroundImage?: string;
  facebookUrl?: string;
  youtubeUrl?: string;
  instagramUrl?: string;
}